fix: correct way to decrement counter
